Clean up all exception output when storing it in the error variable to remove the " at line" stuff added by Perl and the newlines so that the errors stored in objects are consistent.
Fix various bugs in the base object, including a few more type vs. name
inversions and use of object instead of name.  Allow owners to be
specified as ACL names instead of IDs, and change the ID to a name in
show.

Add a new test suite for the base object implementation.