Initial work on supporting testing with MySQL.
Add a drop() method to Wallet::Schema to destroy the wallet database.  Add
a test suite for it.  Add a reinitialize() method to Wallet;:Server that
drops the database before creating it.

Modify the wallet object test cases to call reinitialize() to create the
initial database and drop() to clean up the database after the test is
complete.

Fix a bug preventing Wallet::Schema from being initialized multiple times.
We now stash the schema in a class static variable and reuse it for
subsequent initializations, since re-reading DATA doesn't work.