Provided path to call valid_principal directly
valid_principal used to reside in Wallet::Object::Keytab, but was moved to
the individual Wallet::Kadmin::* modules.  This isn't necessary currently
and may not ever be, but it's there just in case we do ever need to
differentiate.  To simplify testing, a way to still call it directly from
Wallet::Object::Keytab has been added.