Added way to clean things during a fork for kadmin
The MIT kadmin module currently directly runs the MIT kadmin program.  Some
data needs to be cleaned during the forks for this.  This provides a
callback that can be registered and is called during the fork process,
currently just to mark database handles inactive.  It was added to both the
MIT and Heimdal modules, though it's only a stub in the Heimdal module.
Heimdal is not forking kadmin, but the stub is there in order to allow the
caller to not care which module is being used and just always register the
callbacks.