Improved and fixed tests related to Pod and KDC type
Added a fix to the Pod tests to change the order of the arguments in a
skip statement to the correct order.  Also added tests for the
KEYTAB_KRBTYPE value in the keytab tests, and changed the Wallet::Kadmin
module to standardize the errors returned with no keytab set and add new
error for keytab set but not a valid value.