Document and make case-insensitive KEYTAB_KRBTYPE
KEYTAB_KRBTYPE wasn't documented in Wallet::Config.  Add it and the
variable declaration.  Also document the new mandatory setting in
NEWS and add the Heimdal::Kadm5 requirement to README.  Remove some
of the language in README that implies that only MIT Kerberos is
supported.

Make the setting case-insensitive and improve the error message from
Wallet::Kadmin if it isn't set.