Verify that an ACL to be deleted is not referenced
When deleting an ACL on the server, verify that the ACL is not
referenced by any object first.  Database referential integrity should
also catch this, but not all database backends may enforce referential
integrity.  This also allows us to return a better error message
naming an object that's still using that ACL.