Remove initialize and update code out of wallet-admin
wallet-admin is solely a thin wrapper around Wallet::Admin, but it
gained specific code for initialize and update, which caused the
server/admin test to fail.

Move the update code to set a default version into into Wallet::Admin
instead.  The initialize code appears to be unnecessary; it was
setting a default for a parameter that was already handled by
Wallet::Config.

Change-Id: I1a7e5dbbfd005e4f60e89e50a91019295e44df99
Reviewed-on: https://gerrit.stanford.edu/729
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>