Update test configuration documentation
Remove the old tests/data/README documentation and move the note
about krb5.conf into tests/config/README.  (This should change later
to use tests/config for that file.)  Add a license statement to that
file.  Create an empty tests/config directory in the build tree if
the build directory is not the source directory.

Change-Id: Idf466e32cd114841c45f58dba47b11282467ec8f
Reviewed-on: https://gerrit.stanford.edu/848
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>