Improve the keytab enctype test suite
Fix recognition of the syntax error from Heimdal's klist -ke, which
doesn't exit with status 1.  Assume that if we didn't see any known
enctypes, we're dealing with Heimdal.  Remove the code to populate the
enctype table, since we do that in Wallet::Admin now.  Show the error
if adding an enctype fails.

Change-Id: I40da967ef6868e5cd51a1238e1c6324386468134
Reviewed-on: https://gerrit.stanford.edu/1069
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>