Fix wallet-backend parsing of expires
Fix wallet-backend parsing of the expires command to expect only one
argument as the expiration.  This was correctly documented in the
wallet client man page, but not in wallet-backend, and it accepted two
arguments (a date and time).  However, Wallet::Server did not and
would just ignore the time.  Now wallet-backend correctly requires the
date and time be passed as a single argument.

Change-Id: I8e51a576ea8781502f4eb983462ceca867b002be
Reviewed-on: https://gerrit.stanford.edu/1556
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>