Work around Automake distcheck process
Automake insists on not using DESTDIR for distcheck and instead
relying on prefix, but we don't want Perl module installation to
follow prefix since that may result in a module install directory that
isn't in Perl's search path.  So, if and only if we're running under
distcheck, we pass the prefix in as --install_base.

When copying the Test::RRA Perl modules into the perl/t/lib tree,
use separate mkdir and $(INSTALL_DATA) instead of cp -R.  The latter
copies the read-only permissions, and then distclean cannot remove
the files.

Change-Id: Ic1879defad993c76384f7c207cd04cb67889a7ac
Reviewed-on: https://gerrit.stanford.edu/1568
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>