Added a new password object type
The password type inherits almost everything from the file object, but
if you try to get a password object that has never been stored, we
generate a random string to put in the object rather than just
erroring out.  The maximum and minimum length of the string can be set
in the wallet config.

If a password object was stored earlier and then cleared out, we don't
generate another random string.

Change-Id: I17a65ca7dac9d4430e8a731f417297890ee612bb