Implement support for managed Active Directory keytabs
This version implements Active Directory as the store for keytabs.
The interface to Active Directory uses a combination of direct LDAP
queries and the msktutil utility.  This version does not support the
wallet unchanging flag.  Unchanging requires that a keytab be
retrieved without changing the password/kvno which is not supported by
msktutil.