Add error check for partially created AD keytabs
The msktutil script does not always signal error conditions.  This
change implements a check that examines the output from msktutil
and reports and error when the keytab creation fails to create
the keytab but does create a computer entry in the directory.  If
an error is detected the directory entry is deleted leaving the
directory in a clean state.

Also, support has been added for output of debugging information
to syslog using the AD_DEBUG configuration variable.

Finally perltidy suggested changes were made to AD.pm.