Standardize Perl module versions
The versions of all of the wallet Perl modules now match the overall
package version except for Wallet::Schema, which is used to version
the database schema.

Import the test from rra-c-util 5.10 and exclude Wallet::Schema from
the tests.

Go through all Perl modules and standardize the syntax for setting the
version and indicating the required version of Perl.  Fix a few other
syntax issues while I'm in there.