Customize the path to Perl in the server scripts
The server scripts now have a unique string where the Perl path should
be, and the Makefile uses sed to insert the correct path at build time
(using the path determined by configure).

The server scripts now have .in extensions, and setting the executable
flag is handled by the Makefile.

We also have to change autogen, because it looks for the scripts in
the server directory when it generates the man pages.