Clean up generation of Perl scripts
Use @PERL@ as the substitution variable instead of WALLET_PERL_PATH
to match normal Automake conventions.  Write the scripts into the
build path, not the source path, to work properly with out-of-tree
builds.  Clean up the distribution rules.  Use separate Makefile rules
for each script so that make has proper dependencies.

Also fix the Perl paths in (some of) the test scripts, although there
are others for which this still isn't going to work.