Send remctl test output to stderr
We were sending stderr to stdout for remctl output, but sometimes
tests would fail because the remctl output was intermixed with the
test output and confused runtests.  runtests only looks at stdout
for test output and sends stderr to /dev/null, so send remctl
output there.

This isn't a great solution since it interferes with debugging
tests.  The correct solution is to implement the same logging and
polling that's used by C TAP Harness, although that's more difficult
to do in Perl.  But it should make the immediate problem go away.

Continue to send the message about which remctl command line is in
use to stdout, but prefix it with # so that it is a TAP comment.